+++ src/CodingContract/contracts/UniquePathsInAGrid.ts
@@ -1,0 +1,121 @@
+import { getRandomIntInclusive } from "../../utils/helpers/getRandomIntInclusive";
+import { CodingContractTypes } from "../ContractTypes";
+import { CodingContractName } from "@enums";
+
+export const uniquePathsInAGrid: Pick<CodingContractTypes, CodingContractName.UniquePathsInAGridI | CodingContractName.UniquePathsInAGridII> = {
+  [CodingContractName.UniquePathsInAGridI]: {
+    desc: (data: number[]): string => {
+      const numRows = data[0];
+      const numColumns = data[1];
+      return [
+        "You are in a grid with",
+        `${numRows} rows and ${numColumns} columns, and you are`,
+        "positioned in the top-left corner of that grid. You are trying to",
+        "reach the bottom-right corner of the grid, but you can only",
+        "move down or right on each step. Determine how many",
+        "unique paths there are from start to finish.\n\n",
+        "NOTE: The data returned for this contract is an array",
+        "with the number of rows and columns:\n\n",
+        `[${numRows}, ${numColumns}]`,
+      ].join(" ");
+    },
+    difficulty: 3,
+    generate: (): [number, number] => {
+      const numRows: number = getRandomIntInclusive(2, 14);
+      const numColumns: number = getRandomIntInclusive(2, 14);
+
+      return [numRows, numColumns];
+    },
+    solver: (data, answer) => {
+      const n: number = data[0]; // Number of rows
+      const m: number = data[1]; // Number of columns
+      const currentRow: number[] = [];
+      currentRow.length = n;
+
+      for (let i = 0; i < n; i++) {
+        currentRow[i] = 1;
+      }
+      for (let row = 1; row < m; row++) {
+        for (let i = 1; i < n; i++) {
+          currentRow[i] += currentRow[i - 1];
+        }
+      }
+
+      return currentRow[n - 1] === answer;
+    },
+    convertAnswer: (ans) => parseInt(ans, 10),
+    validateAnswer: (ans): ans is number => typeof ans === "number",
+  },
+  [CodingContractName.UniquePathsInAGridII]: {
+    desc: (data: number[][]): string => {
+      let gridString = "";
+      for (const line of data) {
+        gridString += `${line.toString()},\n`;
+      }
+      return [
+        "You are located in the top-left corner of the following grid:\n\n",
+        `${gridString}\n`,
+        "You are trying reach the bottom-right corner of the grid, but you can only",
+        "move down or right on each step. Furthermore, there are obstacles on the grid",
+        "that you cannot move onto. These obstacles are denoted by '1', while empty",
+        "spaces are denoted by 0.\n\n",
+        "Determine how many unique paths there are from start to finish.\n\n",
+        "NOTE: The data returned for this contract is an 2D array of numbers representing the grid.",
+      ].join(" ");
+    },
+    difficulty: 5,
+    generate: (): (1 | 0)[][] => {
+      const numRows: number = getRandomIntInclusive(2, 12);
+      const numColumns: number = getRandomIntInclusive(2, 12);
+
+      const grid: (1 | 0)[][] = [];
+      grid.length = numRows;
+      for (let i = 0; i < numRows; ++i) {
+        grid[i] = [];
+        grid[i].length = numColumns;
+        grid[i].fill(0);
+      }
+
+      for (let r = 0; r < numRows; ++r) {
+        for (let c = 0; c < numColumns; ++c) {
+          if (r === 0 && c === 0) {
+            continue;
+          }
+          if (r === numRows - 1 && c === numColumns - 1) {
+            continue;
+          }
+
+          // 15% chance of an element being an obstacle
+          if (Math.random() < 0.15) {
+            grid[r][c] = 1;
+          }
+        }
+      }
+
+      return grid;
+    },
+    solver: (data, answer) => {
+      const obstacleGrid: number[][] = [];
+      obstacleGrid.length = data.length;
+      for (let i = 0; i < obstacleGrid.length; ++i) {
+        obstacleGrid[i] = data[i].slice();
+      }
+
+      for (let i = 0; i < obstacleGrid.length; i++) {
+        for (let j = 0; j < obstacleGrid[0].length; j++) {
+          if (obstacleGrid[i][j] == 1) {
+            obstacleGrid[i][j] = 0;
+          } else if (i == 0 && j == 0) {
+            obstacleGrid[0][0] = 1;
+          } else {
+            obstacleGrid[i][j] = (i > 0 ? obstacleGrid[i - 1][j] : 0) + (j > 0 ? obstacleGrid[i][j - 1] : 0);
+          }
+        }
+      }
+
+      return obstacleGrid[obstacleGrid.length - 1][obstacleGrid[0].length - 1] === answer;
+    },
+    convertAnswer: (ans) => parseInt(ans, 10),
+    validateAnswer: (ans): ans is number => typeof ans === "number",
+  },
+};
