--- src/CodingContract/contracts/optimization.ts
@@ -1,328 +1,0 @@
-import { CodingContractName } from "@enums";
-import { getRandomIntInclusive } from "../../utils/helpers/getRandomIntInclusive";
-import { type CodingContractTypes } from "../ContractTypes";
-
-export const optimizationDefinitions = {
-  [CodingContractName.SubarrayWithMaximumSum]: {
-    desc: (n: number[]): string => {
-      return [
-        "Given the following integer array, find the contiguous subarray",
-        "(containing at least one number) which has the largest sum and return that sum.",
-        "'Sum' refers to the sum of all the numbers in the subarray.\n",
-        `${n.toString()}`,
-      ].join(" ");
-    },
-    difficulty: 1,
-    generate: (): number[] => {
-      const len: number = getRandomIntInclusive(5, 40);
-      const arr: number[] = [];
-      arr.length = len;
-      for (let i = 0; i < len; ++i) {
-        arr[i] = getRandomIntInclusive(-10, 10);
-      }
-
-      return arr;
-    },
-    solver: (data, answer) => {
-      const nums: number[] = data.slice();
-      for (let i = 1; i < nums.length; i++) {
-        nums[i] = Math.max(nums[i], nums[i] + nums[i - 1]);
-      }
-
-      return Math.max(...nums) === answer;
-    },
-    convertAnswer: (ans) => parseInt(ans, 10),
-    validateAnswer: (ans): ans is number => typeof ans === "number",
-  },
-  [CodingContractName.TotalWaysToSum]: {
-    desc: (n: number): string => {
-      return [
-        "It is possible write four as a sum in exactly four different ways:\n\n",
-        "&nbsp;&nbsp;&nbsp;&nbsp;3 + 1\n",
-        "&nbsp;&nbsp;&nbsp;&nbsp;2 + 2\n",
-        "&nbsp;&nbsp;&nbsp;&nbsp;2 + 1 + 1\n",
-        "&nbsp;&nbsp;&nbsp;&nbsp;1 + 1 + 1 + 1\n\n",
-        `How many different distinct ways can the number ${n} be written as a sum of at least`,
-        "two positive integers?",
-      ].join(" ");
-    },
-    difficulty: 1.5,
-    generate: (): number => {
-      return getRandomIntInclusive(8, 100);
-    },
-    solver: (data, answer) => {
-      if (typeof data !== "number") throw new Error("solver expected number");
-      const ways: number[] = [1];
-      ways.length = data + 1;
-      ways.fill(0, 1);
-      for (let i = 1; i < data; ++i) {
-        for (let j: number = i; j <= data; ++j) {
-          ways[j] += ways[j - i];
-        }
-      }
-
-      return ways[data] === answer;
-    },
-    convertAnswer: (ans) => parseInt(ans, 10),
-    validateAnswer: (ans): ans is number => typeof ans === "number",
-  },
-  [CodingContractName.TotalWaysToSumII]: {
-    desc: (data: [number, number[]]): string => {
-      const n: number = data[0];
-      const s: number[] = data[1];
-      return [
-        `How many different distinct ways can the number ${n} be written`,
-        "as a sum of integers contained in the set:\n\n",
-        `[${s}]?\n\n`,
-        "You may use each integer in the set zero or more times.",
-      ].join(" ");
-    },
-    difficulty: 2,
-    generate: (): [number, number[]] => {
-      const n: number = getRandomIntInclusive(12, 200);
-      const maxLen: number = getRandomIntInclusive(8, 12);
-      const s: number[] = [];
-      // Bias towards small numbers is intentional to have much bigger answers in general
-      // to force people better optimize their solutions
-      for (let i = 1; i <= n; i++) {
-        if (s.length == maxLen) {
-          break;
-        }
-        if (Math.random() < 0.6 || n - i < maxLen - s.length) {
-          s.push(i);
-        }
-      }
-      return [n, s];
-    },
-    solver: (data, answer) => {
-      // https://www.geeksforgeeks.org/coin-change-dp-7/?ref=lbp
-      const n = data[0];
-      const s = data[1];
-      const ways: number[] = [1];
-      ways.length = n + 1;
-      ways.fill(0, 1);
-      for (let i = 0; i < s.length; i++) {
-        for (let j = s[i]; j <= n; j++) {
-          ways[j] += ways[j - s[i]];
-        }
-      }
-      return ways[n] === answer;
-    },
-    convertAnswer: (ans) => parseInt(ans, 10),
-    validateAnswer: (ans): ans is number => typeof ans === "number",
-  },
-  [CodingContractName.MinimumPathSumInATriangle]: {
-    desc: (data: number[][]): string => {
-      function createTriangleRecurse(data: number[][], level = 0): string {
-        const numLevels: number = data.length;
-        if (level >= numLevels) {
-          return "";
-        }
-        const numSpaces = numLevels - level + 1;
-
-        let str: string = ["&nbsp;".repeat(numSpaces), "[", data[level].toString(), "]"].join("");
-        if (level < numLevels - 1) {
-          str += ",";
-        }
-
-        return str + "\n" + createTriangleRecurse(data, level + 1);
-      }
-
-      function createTriangle(data: number[][]): string {
-        return ["[\n", createTriangleRecurse(data), "]"].join("");
-      }
-
-      const triangle = createTriangle(data);
-
-      return [
-        "Given a triangle, find the minimum path sum from top to bottom. In each step",
-        "of the path, you may only move to adjacent numbers in the row below.",
-        "The triangle is represented as a 2D array of numbers:\n\n",
-        `${triangle}\n\n`,
-        "Example: If you are given the following triangle:\n\n[\n",
-        "&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[2],\n",
-        "&nbsp;&nbsp;&nbsp;&nbsp;[3,4],\n",
-        "&nbsp;&nbsp;&nbsp;[6,5,7],\n",
-        "&nbsp;&nbsp;[4,1,8,3]\n",
-        "]\n\n",
-        "The minimum path sum is 11 (2 -> 3 -> 5 -> 1).",
-      ].join(" ");
-    },
-    difficulty: 5,
-    generate: (): number[][] => {
-      const triangle: number[][] = [];
-      const levels: number = getRandomIntInclusive(3, 12);
-      triangle.length = levels;
-
-      for (let row = 0; row < levels; ++row) {
-        triangle[row] = [];
-        triangle[row].length = row + 1;
-        for (let i = 0; i < triangle[row].length; ++i) {
-          triangle[row][i] = getRandomIntInclusive(1, 9);
-        }
-      }
-
-      return triangle;
-    },
-    solver: (data, answer) => {
-      const n: number = data.length;
-      const dp: number[] = data[n - 1].slice();
-      for (let i = n - 2; i > -1; --i) {
-        for (let j = 0; j < data[i].length; ++j) {
-          dp[j] = Math.min(dp[j], dp[j + 1]) + data[i][j];
-        }
-      }
-
-      return dp[0] === answer;
-    },
-    convertAnswer: (ans) => parseInt(ans, 10),
-    validateAnswer: (ans): ans is number => typeof ans === "number",
-  },
-  [CodingContractName.UniquePathsInAGridI]: {
-    desc: (data: number[]): string => {
-      const numRows = data[0];
-      const numColumns = data[1];
-      return [
-        "You are in a grid with",
-        `${numRows} rows and ${numColumns} columns, and you are`,
-        "positioned in the top-left corner of that grid. You are trying to",
-        "reach the bottom-right corner of the grid, but you can only",
-        "move down or right on each step. Determine how many",
-        "unique paths there are from start to finish.\n\n",
-        "NOTE: The data returned for this contract is an array",
-        "with the number of rows and columns:\n\n",
-        `[${numRows}, ${numColumns}]`,
-      ].join(" ");
-    },
-    difficulty: 3,
-    generate: (): [number, number] => {
-      const numRows: number = getRandomIntInclusive(2, 14);
-      const numColumns: number = getRandomIntInclusive(2, 14);
-
-      return [numRows, numColumns];
-    },
-    solver: (data, answer) => {
-      const n: number = data[0]; // Number of rows
-      const m: number = data[1]; // Number of columns
-      const currentRow: number[] = [];
-      currentRow.length = n;
-
-      for (let i = 0; i < n; i++) {
-        currentRow[i] = 1;
-      }
-      for (let row = 1; row < m; row++) {
-        for (let i = 1; i < n; i++) {
-          currentRow[i] += currentRow[i - 1];
-        }
-      }
-
-      return currentRow[n - 1] === answer;
-    },
-    convertAnswer: (ans) => parseInt(ans, 10),
-    validateAnswer: (ans): ans is number => typeof ans === "number",
-  },
-  [CodingContractName.ArrayJumpingGame]: {
-    desc: (arr: number[]): string => {
-      return [
-        "You are given the following array of integers:\n\n",
-        `${arr}\n\n`,
-        "Each element in the array represents your MAXIMUM jump length",
-        "at that position. This means that if you are at position i and your",
-        "maximum jump length is n, you can jump to any position from",
-        "i to i+n.",
-        "\n\nAssuming you are initially positioned",
-        "at the start of the array, determine whether you are",
-        "able to reach the last index.\n\n",
-        "Your answer should be submitted as 1 or 0, representing true and false respectively.",
-      ].join(" ");
-    },
-    difficulty: 2.5,
-    generate: (): number[] => {
-      const len: number = getRandomIntInclusive(3, 25);
-      const arr: number[] = [];
-      arr.length = len;
-      for (let i = 0; i < arr.length; ++i) {
-        if (Math.random() < 0.2) {
-          arr[i] = 0; // 20% chance of being 0
-        } else {
-          arr[i] = getRandomIntInclusive(0, 10);
-        }
-      }
-
-      return arr;
-    },
-    numTries: 1,
-    solver: (data, answer) => {
-      const n: number = data.length;
-      let i = 0;
-      for (let reach = 0; i < n && i <= reach; ++i) {
-        reach = Math.max(i + data[i], reach);
-      }
-      const solution: boolean = i === n;
-      return (solution ? 1 : 0) === answer;
-    },
-    convertAnswer: (ans) => {
-      const num = parseInt(ans);
-      if (num === 0 || num === 1) return num;
-      return null;
-    },
-    validateAnswer: (ans): ans is 1 | 0 => typeof ans === "number" && (ans === 0 || ans === 1),
-  },
-  [CodingContractName.ArrayJumpingGameII]: {
-    desc: (arr: number[]): string => {
-      return [
-        "You are given the following array of integers:\n\n",
-        `${arr}\n\n`,
-        "Each element in the array represents your MAXIMUM jump length",
-        "at that position. This means that if you are at position i and your",
-        "maximum jump length is n, you can jump to any position from",
-        "i to i+n.",
-        "\n\nAssuming you are initially positioned",
-        "at the start of the array, determine the minimum number of",
-        "jumps to reach the end of the array.\n\n",
-        "If it's impossible to reach the end, then the answer should be 0.",
-      ].join(" ");
-    },
-    difficulty: 3,
-    generate: (): number[] => {
-      const len: number = getRandomIntInclusive(3, 25);
-      const arr: number[] = [];
-      arr.length = len;
-      for (let i = 0; i < arr.length; i++) {
-        for (let j = 0; j < 10; j++) {
-          if (Math.random() <= j / 10 + 0.1) {
-            arr[i] = j;
-            break;
-          }
-        }
-      }
-
-      return arr;
-    },
-    numTries: 3,
-    solver: (data, answer) => {
-      const n: number = data.length;
-      let reach = 0;
-      let jumps = 0;
-      let lastJump = -1;
-      while (reach < n - 1) {
-        let jumpedFrom = -1;
-        for (let i = reach; i > lastJump; i--) {
-          if (i + data[i] > reach) {
-            reach = i + data[i];
-            jumpedFrom = i;
-          }
-        }
-        if (jumpedFrom === -1) {
-          jumps = 0;
-          break;
-        }
-        lastJump = jumpedFrom;
-        jumps++;
-      }
-      return jumps === answer;
-    },
-    convertAnswer: (ans) => parseInt(ans, 10),
-    validateAnswer: (ans): ans is number => typeof ans === "number",
-  },
-} satisfies Partial<CodingContractTypes>;
